Remove incorrect _update_pits
This call to _update_pits was there to update the pits after the handling of last stone in empty pit was done. However, it was incorrect to place it there because _handle_last_in_empty already updated the pits, all the call for _update_pits after it basically deleted what _handle_last_in_empty did, because it used the all_pits list, which _handle_last_in_empty does not update.